Parse NT as real integer.

+++ trunk/1phase.c
@@ -26,7 +26,7 @@ static long double DX;
 /* uniform timesteps */
 static long double DT;
 /* number of timesteps */
-static long double NT;
+static long NT;
 
 #define print_array(a) __print_array(#a, a)
 
@@ -59,7 +59,7 @@ static struct param params[] = {
 	}, {
 		.name = "NT",
 		.value = &NT,
-		.type = FLOAT,
+		.type = INTEGER,
 	}, {
 		.name = "K",
 		.value = &K,
@@ -174,8 +174,8 @@ int main(void)
 		die("%s: DX = %Lf <= 0.0", __func__, DX);
 	if (DT <= 0.0)
 		die("%s: DT = %Lf <= 0.0", __func__, DT);
-	if ((int)NT <= 0)
-		die("%s: NT = %d <= 0", __func__, (int)NT);
+	if (NT <= 0)
+		die("%s: NT = %d <= 0", __func__, NT);
 	if (K <= 0.0)
 		die("%s: K = %Lf <= 0.0", __func__, K);
 	if (phi <= 0.0 || phi > 1.0)
@@ -189,7 +189,7 @@ int main(void)
 		p[i] = p_x0(i);
 	print_array(p);
 
-	for (t = 1; t < (int)NT; t++) {
+	for (t = 1; t < NT; t++) {
 		p1[0] = p_lt(t);
 		for (i = 1; i <= N - 2; i++)
 			p1[i] = p[i] + DT * K / (phi * mu * c) * d2f_dx2(i, p);
